+++ testutil/integration/base/grpc/grpc.go
@@ -5,7 +5,6 @@ import (
 	precisebanktypes "github.com/cosmos/evm/x/precisebank/types"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/authz"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	distrtypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
@@ -16,14 +15,6 @@ import (
 type Handler interface {
 	// Account methods
 	GetAccount(address string) (sdk.AccountI, error)
-
-	// Authz methods
-	GetAuthorizations(grantee, granter string) ([]authz.Authorization, error)
-	GetAuthorizationsByGrantee(grantee string) ([]authz.Authorization, error)
-	GetAuthorizationsByGranter(granter string) ([]authz.Authorization, error)
-	GetGrants(grantee, granter string) ([]*authz.Grant, error)
-	GetGrantsByGrantee(grantee string) ([]*authz.GrantAuthorization, error)
-	GetGrantsByGranter(granter string) ([]*authz.GrantAuthorization, error)
 
 	// Bank methods
 	GetBalanceFromBank(address sdk.AccAddress, denom string) (*banktypes.QueryBalanceResponse, error)